+++ ricaun.Nuke/Components/IHazSolution.cs
@@ -8,4 +8,14 @@ namespace ricaun.Nuke.Components
     {
         [Required] [Solution] Solution Solution => ValueInjectionUtility.TryGetValue(() => Solution);
     }
+
+    public static class HazSolutionExtension
+    {
+        /// <summary>
+        /// Get Solution
+        /// </summary>
+        /// <param name="hazSolution"></param>
+        /// <returns></returns>
+        public static Solution GetMainProject(this IHazSolution hazSolution) => hazSolution.Solution;
+    }
 }