+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.0.10] / 2021-12-17
+- Add GetMainProject on IHazMainProject 
+- private GetMainProject
+- Add HazSolutionExtension
+- Add HazMainProjectExtension
+- Add IHazMainProject
 - Rename Example Folder Project
 
 ## [0.0.9] / 2021-12-17